Separate setup macros and rules across .bzl files
This doesn't yet resolve the problem from the previous commit, but
refines the solution from the previous two commits.

The commits importing `java_common` and `JavaInfo` fixed the same
problem as this commit, but left the `@rules_java` dependency in the
loading phase. It's best to partition these dependencies and the rules
that use them from loading phase code altogether, as done in this
commit.

Per the following, it's good practice to ensure that the transitive
closure of `.bzl` files containing setup macros contains no `BUILD`
macros, rules, etc.:

- https://github.com/bazel-contrib/rules_scala/issues/1652#issuecomment-2507453047
- https://blog.engflow.com/2025/07/08/migrating-to-bazel-modules-aka-bzlmod---maintaining-compatibility-part-1/#separate-configuration-bzl-files-from-rule-and-aspect-bzl-files

While these intermingled dependencies worked under Bazel 6, 7, and 8,
support is going away in Bazel 9. In fact, before the changes in the
previous two commits, Bazel 8.0.0rc6 also failed due to a now missing
`java_common` symbol.

Due to the discussion in the above `rules_scala` issue, Bazel 8 actually
provides `java_common` stubs to maintain compatibility in such cases,
thanks to:

- https://github.com/bazelbuild/bazel/commit/15b8e3f7e07e1d286fa73e73c55da45e7508d73d

But this support is gone again as of Bazel 9.0.0-pre.20250516.1:

- https://github.com/bazelbuild/bazel/commit/961b83950acbf557ed94c68a05761c2b5ec0a2ae