Use the Bash runfiles library in magic_test
As an alternative to the original implementation, the `data` attribute
is now an `attr.label_list`. This enables the test implementation to use
`$(rlocation)` with a repo-relative path, instead of passing in a
`$(rootpath)` manually from a rule attribute. It works with Bazel 6, 7,
and 8.

Technically, we could pass `@io_frobozzco_rules_bootcamp_leaflet` to
`deps` instead of `data`. It doesn't really matter, but using `data` is
more canonical for runfile dependencies that aren't required to build
the target.