Import as rules_magic, not com_frobozz_rules_magic
This demonstrates why target strings that contain a reference to their
own repo name are a liability. It forces `http_archive` users to import
the repo using the same `name` it gives itself. (It may not affect
`bazel_dep` users, due to Bzlmod's repo mapping mechanism, but as we'll
see, it's still unnecessary.)

When users use a different name (perhaps a shorter one), it causes that
repo to break due to an apparent repository reference cycle. This is
because there's no repo with the expected name, so a `load` statement
can't find its file:

```txt
ERROR: Failed to load Starlark extension
  '@@com_frobozz_rules_magic//magic:private/setup_magic_toolchain.bzl'.

Cycle in the workspace file detected. This indicates that a repository
is used prior to being defined.

The following chain of repository dependencies lead to the missing
definition.
 - @@com_frobozz_rules_magic

This could either mean you have to add the '@@com_frobozz_rules_magic'
repository with a statement like `http_archive` in your WORKSPACE file
(note that transitive dependencies are not added automatically), or move
an existing definition earlier in your WORKSPACE file.

ERROR: Error computing the main repository mapping:
  cycles detected during computation of main repo mapping
```

Or, worse, Bazel instantiates an unexpected version of the repo as a
transitive dependency. The build may continue, but with two versions of
the same repo, which may or may not produce an error (which will likely
prove confusing).