Import as rules_magic, not com_frobozz_rules_magic
This demonstrates why target strings that contain a reference to their
own repo name are a liability. It forces users to import the repo using
the same name it gives itself. When users use a different name (perhaps
a shorter one), it causes that repo to break due to an apparent
repository reference cycle:

```txt
ERROR: Failed to load Starlark extension
  '@@com_frobozz_rules_magic//magic:toolchains.bzl'.

Cycle in the workspace file detected. This indicates that a repository
is used prior to being defined.

The following chain of repository dependencies lead to the missing
definition.
 - @@com_frobozz_rules_magic

This could either mean you have to add the '@@com_frobozz_rules_magic'
repository with a statement like `http_archive` in your WORKSPACE file
(note that transitive dependencies are not added automatically), or move
an existing definition earlier in your WORKSPACE file.

ERROR: Error computing the main repository mapping:
  cycles detected during computation of main repo mapping
```