Bump to protobuf v29.0
This is the minimum `protobuf` version required by Bazel 8. However, the
build breaks the Bazel 8 build with:

```txt
ERROR: Failed to load Starlark extension '@@bazel_features//:deps.bzl'.
Cycle in the workspace file detected. This indicates that a repository
is used prior to being defined.

The following chain of repository dependencies lead to the missing
definition.
 - @@bazel_features

This could either mean you have to add the '@@bazel_features' repository
with a statement like `http_archive` in your WORKSPACE file (note that
transitive dependencies are not added automatically), or move an
existing definition earlier in your WORKSPACE file.

ERROR: Error computing the main repository mapping: cycles detected
during computation of main repo mapping
```

This is because Bazel 8.3.0 updated the legacy WORKSPACE suffix to load
`@bazel_features//:deps.bzl`, but doesn't instantiate the
@bazel_features` repo. No other existing dependency happens to
instantiate `@bazel_features`, either.

- bazelbuild/bazel@0e528d037e041ea754e4420de6b4c9f2e502b57e
- https://github.com/bazelbuild/bazel/blob/8.4.2/src/main/java/com/google/devtools/build/lib/bazel/rules/BUILD#L136-L137

---

The Bazel 7 build also breaks with this change. After the next commit to
add the `bazel_features` repo, the Bazel 8 build will break in the same
way, so we'll describe it then.